+ dictionary to array for a fast lookup + temporary removed the dynamic sound. Often soudns are not played when recalling

+++ src/Network/PacketHandlers.cs
@@ -1673,7 +1673,7 @@ namespace ClassicUO.Network
             ushort audio = p.ReadUShort();
             ushort x = p.ReadUShort();
             ushort y = p.ReadUShort();
-            ushort z = p.ReadUShort();
+            short z = (short) p.ReadUShort();
 
             Client.Game.Scene.Audio.PlaySoundWithDistance(index, x, y);
         }
